todo excepto Log y Sign reacomodado
