+++ client/src/App.js
@@ -1,6 +1,6 @@
 import {Route} from 'react-router-dom'
 import Nav from './components/Header/Nav';
-import Dashboard from './components/Admin/Professional/Dashboard';
+import Dashboard from './components/Professional/ProfessionalDashboard';
 import PatientsDashboard from './components/Patient/PatientsDashboard';
 import Sign from './components/Sign/Sign';
 import Log from './components/Log/Log';
