+++ client/src/components/Header/Buttons/ButtonLogin.js
@@ -1,5 +1,5 @@
 import Button from '@material-ui/core/Button'
-import { makeStyles,Typography,Box} from '@material-ui/core';
+import { makeStyles} from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
 const useStyle=makeStyles(theme=>({
     btn:{
@@ -9,15 +9,15 @@ const useStyle=makeStyles(theme=>({
     },
 }))
 
-export default function ButtonSignIn(){
+export default function ButtonLogin(){
     let history=useHistory();
     const classes=useStyle();
     function handleClick(btnlink){
         history.push(btnlink)
     }
     return(
-        <Button type="button" variant='contained' className={classes.btn} onClick={()=>handleClick("/sign-In")}> 
-            Regístrate
+        <Button type="button" variant='contained' className={classes.btn} onClick={()=>handleClick("/login")}>
+            Iniciar sesión
         </Button>
     )
 }
