+++ client/src/components/Header/Test/Nav.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect'
 import {render} from '@testing-library/react'
-import Nav from './nav';
+import Nav from '../nav';
 
 test('render nav', ()=>{
     const component = render(<Nav/>)
