+++ client/src/components/Main/Carousel/Carousel.test.js
