+++ client/src/components/Patient/NavPanel/NavPanel.js
@@ -1,17 +1,13 @@
 import {
-  Container,
-  Grid,
   Box,
-  Link,
   Button,
-  Typography,
   makeStyles,
 } from "@material-ui/core";
-import { useEffect, useState } from "react";
-import Appointment from "../Appointment";
-import Attention from "../Attention";
-import Professionals from "../Professionals";
-import Managment from "../Managment";
+import { useEffect} from "react";
+import Appointment from "../SubComponents/Appointment";
+import Attention from "../SubComponents/Attention";
+import Professionals from "../SubComponents/Professionals";
+import Managment from "../SubComponents/Managment";
 
 const useStyle = makeStyles((theme) => ({
   magin: {
@@ -57,9 +53,11 @@ const useStyle = makeStyles((theme) => ({
   },
 }));
 
+//Administra el componente a renderizar y muestra los botones y opciones
 export default function NavPanel({ updateComponent }) {
   const classes = useStyle();
-  const [actComp, setaActComp] = useState("Inicio");
+
+//Objeto de componentes y nombres
   const routes = [
     { Appointment: <Appointment /> },
     { Managment: <Managment /> },
