+++ client/src/components/Professional/SubComponents/Stock/StockCard/StockCard.js
